Adds PHPUnit 10 support
