+++ src/Illuminate/Testing/Assert.php
@@ -39,41 +39,4 @@ abstract class Assert extends PHPUnit
 
         PHPUnit::assertThat($array, $constraint, $msg);
     }
-
-    /**
-     * Asserts that a file does not exist.
-     *
-     * @param  string  $filename
-     * @param  string  $message
-     * @return void
-     */
-    public static function assertFileDoesNotExist(string $filename, string $message = ''): void
-    {
-        static::assertThat($filename, new LogicalNot(new FileExists), $message);
-    }
-
-    /**
-     * Asserts that a directory does not exist.
-     *
-     * @param  string  $directory
-     * @param  string  $message
-     * @return void
-     */
-    public static function assertDirectoryDoesNotExist(string $directory, string $message = ''): void
-    {
-        static::assertThat($directory, new LogicalNot(new DirectoryExists), $message);
-    }
-
-    /**
-     * Asserts that a string matches a given regular expression.
-     *
-     * @param  string  $pattern
-     * @param  string  $string
-     * @param  string  $message
-     * @return void
-     */
-    public static function assertMatchesRegularExpression(string $pattern, string $string, string $message = ''): void
-    {
-        static::assertThat($string, new RegularExpression($pattern), $message);
-    }
 }
